# perf: parallelize ngram indexing